Fix missing table columns of creation of online_transaction

+++ src/main/java/com/mvidyn/std/phub/ui/repository/ApplicantRepository.java
@@ -3,5 +3,5 @@ package com.mvidyn.std.phub.ui.repository;
 import com.mvidyn.std.phub.ui.model.Applicant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicantRepository extends JpaRepository<Applicant, Integer> {
+public interface ApplicantRepository extends JpaRepository<Applicant, Long> {
 }
