+++ src/main/java/com/mvidyn/std/phub/ui/repository/BeneficiaryRepository.java
@@ -3,5 +3,5 @@ package com.mvidyn.std.phub.ui.repository;
 import com.mvidyn.std.phub.ui.model.Beneficiary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BeneficiaryRepository extends JpaRepository<Beneficiary, Integer> {
+public interface BeneficiaryRepository extends JpaRepository<Beneficiary, Long> {
 }
